adds suport for actually de-sugaring a grammar
implements the getTransformedGrammar method
of SugarTransformer, adds a builder to the
Rule class, adds an alternate constructor
to the grammar class, and some copy methods
for rules and RHSTree objects.

creates a static method on the Grammar class for
the creation of the default grammar tokenizer,
and uses it everywhere instead of using the new
keyword.

Updates the SugarTest to test the de-sugarization
of the grammars, as well as that the new grammar
is the same as the old (up to some threshold)