implements removing left recursion via LeftRecursionTransformer
a couple of small changes to Grammar and SugarTransformer.

adds more testing to LeftRecursiveTest, as well as changes to
simpleIndirect.ebnf to differentiate it from simpleDirect.ebnf

finishes up LeftRecursionTransformer using the algorithm from
the paper. However, there is a small change to case 1 to prevent
non-terminals showing up on the RHS without definitions.