more LR testing and a regex fix
for the LR test cases that can be described by
a regex, adds a test to compare the transformed
grammar to that regex.

This uncovered an issue with the RegexExpecterator,
in that it wasn't allowed to consume more than one
token (so something like 'a+' would only match a
single 'a' but not match 'aa'). This is now fixed,
but more testing might be a good idea in this area.