add calculator example
creates ParseResultFunction as a functional interface
so that it can throw a ParseResultException instead of
just being a Function<ParseResult, ParseResult>.

This can be useful for parsing something like "1 / 0"
which is syntactically valid but not semantically valid.
Now ParseResultTransformer::transform can catch that
exception when trying to apply the ParseResultFunction.

creates a context free language for a simple calculator.
adds a unit test, TestCalculator, to test some expressions
using the calculator as the grammar in a parser.

fleshes out some numeric parse result classes.