adds regex rule test
adds a test to ParserTest for testing some regex rules.
Ends up changing Tokenizer::tokenize so that it can throw
an exception of type TokenizerException instead of throwing
a RuntimeException so that the exception can be caught in
the Parser::parseString method. This caused a cascade of
changes, and I think I should make even more changes so
that exceptions are handled in a better way all over the
project. That can come later though.